feat: decentralized encoder architecture and lazy registration
